tests(data-fbs): refactor flatbuffers tests
This commit includes the following:
 - Refactor: flatbuffer tests now are in its own folder
 - Refactor: test was spllited into 2 different tests (encode and decode)
 - Fix: lint error
